added week 4
We start with a brief review of linear algrebra, and then extend the topic to convolutions using audio data as example. key concepts like locality, stationality and Toeplitz matrix are discussed. Then we give a live demo of convolution performance in pitch analysis. Finally there is a short diagression about the dimensionality of different data.