Update Image 4(a) and addition of Image 4(b)
Please check is the formatting for markdown looks good on the website. I cannot correctly preview changes from browser with the Preview buttom.

Also, move images 4(a) and 4(b) to the repo "./images/" folder , I don 't have privileges to do that (to the best of my knowledge), so I've used  random URLs that Github generated to store images.